ADD: add userdefaults to store volume settings.
